requirements.txt
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1